src/job_api.py
from apify_client import ApifyClient
import os
from dotenv import load_dotenv
load_dotenv()

apify_client=ApifyClient(os.getenv("APIFY_API_TOKEN")) 

#fetch linkedin jobs using apify
def fetch_linkedin_jobs(search_query, location="india", rows=60):
    """Fetch job listings from LinkedIn using Apify."""
    run_input = {
        "title": search_query,
        "location": location,
        "rows": rows,
        "proxy": {
            "useApifyProxy": True,
            "apifyProxyGroups": ["RESIDENTIAL"],
        },
    }
    run = apify_client.actor("BHzefUZlZRKWxkTck").call(run_input=run_input)
    jobs=list(apify_client.dataset(run["defaultDatasetId"]).iterate_items())
    return jobs

#fetch naukri jobs using apify
def fetch_naukri_jobs(search_query, location="india", rows=60):
    """Fetch job listings from Naukri using Apify."""
    run_input = {
        "keywords": search_query,
        "maxJobs": rows,
        "freshness": "all",
        "sortBy": "relevance",
        "experience":"all",
    }

    # Run the Actor and wait for it to finish
    run = apify_client.actor("wsrn5gy5C4EDeYCcD").call(run_input=run_input)
    jobs=list(apify_client.dataset(run["defaultDatasetId"]).iterate_items())
    return jobs
